refactor(it): split caddy + network/status under the 500-line structure cap
CI's Structure check (file-lengths ≤500) failed: caddy.rs had grown to 572
lines and network/status.rs to 670. Both `it/` and `it/network/` are already
at the 8-entry folder cap, so a sibling file is not an option — extract each
file's `#[cfg(test)] mod tests` into a `tests.rs` under a new module directory:

  it/caddy.rs           -> it/caddy/{mod.rs (366), tests.rs (205)}
  it/network/status.rs  -> it/network/status/{mod.rs (479), tests.rs (190)}

Pure move: tests keep `use super::*` (super = the module), no visibility
changes, no behaviour change. Parent folders stay at 8 entries; the new dirs
hold 2 files each. Neither file is in the protected-files hash chain.

Verified: cargo fmt --check clean, all 30 moved tests pass (14 caddy + 16
status), crate compiles, structure check passes on tracked files.
